Refactor code to duduplicate logic of creating order documents
