+++ src/Core/Checkout/Document/DocumentGenerator/CreditNoteGenerator.php
@@ -3,50 +3,21 @@
 namespace Shopware\Core\Checkout\Document\DocumentGenerator;
 
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
-use Shopware\Core\Checkout\Document\DocumentConfiguration;
-use Shopware\Core\Checkout\Document\DocumentConfigurationFactory;
-use Shopware\Core\Checkout\Document\Twig\DocumentTemplateRenderer;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Shopware\Core\Framework\Context;
-use Twig\Error\Error;
 
-class CreditNoteGenerator implements DocumentGeneratorInterface
+class CreditNoteGenerator extends AbstractOrderDocumentGenerator implements DocumentGeneratorInterface
 {
-    public const DEFAULT_TEMPLATE = '@Framework/documents/credit_note.html.twig';
     public const CREDIT_NOTE = 'credit_note';
-
-    /**
-     * @var string
-     */
-    private $rootDir;
-
-    /**
-     * @var DocumentTemplateRenderer
-     */
-    private $documentTemplateRenderer;
-
-    public function __construct(DocumentTemplateRenderer $documentTemplateRenderer, string $rootDir)
-    {
-        $this->documentTemplateRenderer = $documentTemplateRenderer;
-        $this->rootDir = $rootDir;
-    }
+    public const DEFAULT_TEMPLATE = '@Framework/documents/credit_note.html.twig';
 
     public function supports(string $documentType): bool
     {
         return $documentType === self::CREDIT_NOTE;
     }
 
-    /**
-     * @throws Error
-     */
-    public function generate(
-        OrderEntity $order,
-        DocumentConfiguration $config,
-        Context $context,
-        ?string $templatePath = null
-    ): string {
-        $templatePath = $templatePath ?? self::DEFAULT_TEMPLATE;
-
+    protected function getExtraParameters(OrderEntity $order, Context $context): array
+    {
         $lineItems = $order->getLineItems();
         $creditItems = [];
         if ($lineItems) {
@@ -57,26 +28,13 @@ class CreditNoteGenerator implements DocumentGeneratorInterface
             }
         }
 
-        $documentString = $this->documentTemplateRenderer->render(
-            $templatePath,
-            [
-                'order' => $order,
-                'creditItems' => $creditItems,
-                'config' => DocumentConfigurationFactory::mergeConfiguration($config, new DocumentConfiguration())->jsonSerialize(),
-                'rootDir' => $this->rootDir,
-                'context' => $context,
-            ],
-            $context,
-            $order->getSalesChannelId(),
-            $order->getLanguageId(),
-            $order->getLanguage()->getLocale()->getCode()
-        );
-
-        return $documentString;
+        return [
+            'creditItems' => $creditItems,
+        ];
     }
 
-    public function getFileName(DocumentConfiguration $config): string
+    protected function getDefaultTemplate(): string
     {
-        return $config->getFilenamePrefix() . $config->getDocumentNumber() . $config->getFilenameSuffix();
+        return self::DEFAULT_TEMPLATE;
     }
 }
