+++ src/Core/Checkout/Document/DocumentGenerator/InvoiceGenerator.php
@@ -2,69 +2,26 @@
 
 namespace Shopware\Core\Checkout\Document\DocumentGenerator;
 
-use Shopware\Core\Checkout\Document\DocumentConfiguration;
-use Shopware\Core\Checkout\Document\DocumentConfigurationFactory;
-use Shopware\Core\Checkout\Document\Twig\DocumentTemplateRenderer;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Shopware\Core\Framework\Context;
-use Twig\Error\Error;
 
-class InvoiceGenerator implements DocumentGeneratorInterface
+class InvoiceGenerator extends AbstractOrderDocumentGenerator implements DocumentGeneratorInterface
 {
-    public const DEFAULT_TEMPLATE = '@Framework/documents/invoice.html.twig';
     public const INVOICE = 'invoice';
-
-    /**
-     * @var string
-     */
-    private $rootDir;
-
-    /**
-     * @var DocumentTemplateRenderer
-     */
-    private $documentTemplateRenderer;
-
-    public function __construct(DocumentTemplateRenderer $documentTemplateRenderer, string $rootDir)
-    {
-        $this->rootDir = $rootDir;
-        $this->documentTemplateRenderer = $documentTemplateRenderer;
-    }
+    public const DEFAULT_TEMPLATE = '@Framework/documents/invoice.html.twig';
 
     public function supports(string $documentType): bool
     {
         return $documentType === self::INVOICE;
     }
 
-    /**
-     * @throws Error
-     */
-    public function generate(
-        OrderEntity $order,
-        DocumentConfiguration $config,
-        Context $context,
-        ?string $templatePath = null
-    ): string {
-        $templatePath = $templatePath ?? self::DEFAULT_TEMPLATE;
-
-        $documentString = $this->documentTemplateRenderer->render(
-            $templatePath,
-            [
-                'order' => $order,
-                'config' => DocumentConfigurationFactory::mergeConfiguration($config, new DocumentConfiguration())->jsonSerialize(),
-                'rootDir' => $this->rootDir,
-                'context' => $context,
-            ],
-            $context,
-            $order->getSalesChannelId(),
-            $order->getLanguageId(),
-            $order->getLanguage()->getLocale()->getCode()
-        );
-
-        return $documentString;
+    protected function getExtraParameters(OrderEntity $order, Context $context): array
+    {
+        return [];
     }
 
-    public function getFileName(DocumentConfiguration $config): string
+    protected function getDefaultTemplate(): string
     {
-        return $config->getFilenamePrefix() . $config->getDocumentNumber() . $config->getFilenameSuffix();
+        return self::DEFAULT_TEMPLATE;
     }
 }
